removed external api

+++ app/src/main/java/com/jetpack/carpartsfinder/dto/ImageData.kt
@@ -3,5 +3,5 @@ package com.jetpack.carpartsfinder.dto
 data class ImageData(
     val id: String,
     val path: String,
-    val rating: Float?,
+    val rating: Short?,
 )
